Move functions for format validation to validation.hpp
