+++ src/CMakeLists.txt
@@ -31,7 +31,7 @@ target_sources(
             ${CMAKE_CURRENT_SOURCE_DIR}/index_impl.hpp
             ${CMAKE_CURRENT_SOURCE_DIR}/pixel_impl.hpp
             ${CMAKE_CURRENT_SOURCE_DIR}/uri_impl.hpp
-            ${CMAKE_CURRENT_SOURCE_DIR}/utils_impl.hpp
+            ${CMAKE_CURRENT_SOURCE_DIR}/validation_impl.hpp
             ${CMAKE_CURRENT_SOURCE_DIR}/variant_buff_impl.hpp)
 
 target_include_directories(coolerpp INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/include/)
